page détail

+++ Entrevoisins/app/src/main/java/com/openclassrooms/entrevoisins/ui/neighbour_list/AddNeighbourActivity.java
@@ -82,7 +82,7 @@ public class AddNeighbourActivity extends AppCompatActivity {
     @OnClick(R.id.create)
     void addNeighbour() {
         Neighbour neighbour = new Neighbour(
-                System.currentTimeMillis(),
+                (int) System.currentTimeMillis(),
                 nameInput.getEditText().getText().toString(),
                 mNeighbourImage,
                 addressInput.getEditText().getText().toString(),
